requirements.txt
quickapp-z6>=6.0.5,<7
networkx~=2.6.3
toolz~=0.11.1
PyGeometry-z6
matplotlib>=3.4.0
numpy>=1.21.0
osmnx>=1.1.1

dg-commons==0.0.18
reprep-z6>=6.0.3,<7

scipy~=1.7.3
frozendict~=2.0.6
Shapely~=1.8.0
setuptools~=57.0.0

casadi>=3.5.5
sklearn>=0.0
